--- test_runner/src/main/kotlin/ftl/domain/RunDoctorIos.kt
@@ -1,20 +1,0 @@
-package ftl.domain
-
-import ftl.args.IosArgs
-import ftl.doctor.validateYaml
-import ftl.presentation.Output
-import ftl.presentation.cli.firebase.test.processValidation
-import ftl.run.exception.YmlValidationError
-import java.nio.file.Paths
-
-interface RunDoctorIos : Output {
-    val configPath: String
-    val fix: Boolean
-}
-
-operator fun RunDoctorIos.invoke() {
-    val validationResult = validateYaml(IosArgs, Paths.get(configPath))
-    if (fix) processValidation(Paths.get(configPath))
-    validationResult.out()
-    if (!validationResult.isEmpty()) throw YmlValidationError()
-}
