+++ test_runner/src/main/kotlin/ftl/presentation/cli/firebase/test/CommandUtil.kt
@@ -1,8 +1,13 @@
 package ftl.presentation.cli.firebase.test
 
 import flank.common.logLn
+import ftl.args.yml.DEVICES_NODE
+import ftl.args.yml.GCLOUD_NODE
+import ftl.args.yml.VERSION_NODE
 import ftl.args.yml.YamlDeprecated
 import ftl.args.yml.fixDevices
+import ftl.domain.RunDoctor
+import ftl.domain.isEmpty
 import ftl.run.exception.YmlValidationError
 import java.nio.file.Path
 
@@ -13,3 +18,16 @@ fun processValidation(ymlPath: Path) {
     }
     fixDevices(ymlPath)
 }
+
+fun RunDoctor.Error.summary(): String =
+    if (isEmpty()) "Valid yml file"
+    else buildString {
+        parsingErrors.forEach { appendLine(it) }
+        invalidDevices
+            .map { "Warning: Version should be string $GCLOUD_NODE -> $DEVICES_NODE[${it.model}] -> $VERSION_NODE[${it.version}]" }
+            .forEach { appendLine(it) }
+        if (topLevelUnknownKeys.isNotEmpty()) appendLine("Unknown top level keys: $topLevelUnknownKeys")
+        nestedUnknownKeys.forEach { (topLevelKey, keys) ->
+            appendLine("Unknown keys in $topLevelKey -> $keys")
+        }
+    }.trim()
