+++ test_runner/src/main/kotlin/ftl/presentation/cli/firebase/test/ios/IosDoctorCommand.kt
@@ -1,10 +1,10 @@
 package ftl.presentation.cli.firebase.test.ios
 
+import ftl.args.IosArgs
 import ftl.config.FtlConstants
-import ftl.domain.DoctorErrors
-import ftl.domain.RunDoctorIos
+import ftl.domain.RunDoctor
 import ftl.domain.invoke
-import ftl.domain.summary
+import ftl.presentation.cli.firebase.test.summary
 import ftl.presentation.outputLogger
 import ftl.presentation.throwUnknownType
 import picocli.CommandLine.Command
@@ -27,7 +27,7 @@ import picocli.CommandLine.Option
 )
 class IosDoctorCommand :
     Runnable,
-    RunDoctorIos {
+    RunDoctor {
 
     @Option(
         names = ["-h", "--help"],
@@ -48,11 +48,11 @@ class IosDoctorCommand :
     )
     override var fix: Boolean = false
 
-    override fun run() = invoke()
+    override fun run() = invoke(IosArgs)
 
     override val out = outputLogger {
         when (this) {
-            is DoctorErrors -> summary()
+            is RunDoctor.Error -> summary()
             else -> throwUnknownType()
         }
     }
